Пути контроллеров приведены к требованиям RESTAPI

+++ src/main/java/ru/solarlab/study/repository/AdvertisementRepository.java
@@ -4,10 +4,8 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 import ru.solarlab.study.entity.Advertisement;
 
-import java.util.List;
-
 @Repository
-public interface AdvertisementRepository extends PagingAndSortingRepository<Advertisement, Integer> {
+public interface AdvertisementRepository extends PagingAndSortingRepository<Advertisement, Long> {
 
     //List<Advertisement> findById(Integer id);
