+++ src/main/java/ru/solarlab/study/repository/CategoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import ru.solarlab.study.entity.Category;
 
 @Repository
-public interface CategoryRepository extends PagingAndSortingRepository<Category, Integer> {
+public interface CategoryRepository extends PagingAndSortingRepository<Category, Long> {
 
     //List<Category> findById(Integer id);
