+++ src/main/java/ru/solarlab/study/repository/TagRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import ru.solarlab.study.entity.Tag;
 
 @Repository
-public interface TagRepository extends PagingAndSortingRepository<Tag, Integer> {
+public interface TagRepository extends PagingAndSortingRepository<Tag, Long> {
 
     //List<Category> findById(Integer id);
